Refactor DnD to a composable.
Refs #74.
